Extracted out CollectionTask.SaveModel to saveModel
